2003-05-25  Christian Stimming  <stimming@tuhh.de>
	* src/gnome-utils/gnc-tree-model-example-account.c,
	src/backend/file/sixtp-dom-parsers.c,
	src/backend/file/sixtp-utils.c,
	src/backend/file/test/test-dom-converters1.c,
	src/backend/file/test/test-string-converters.c,
	src/calculation/fin.c, src/engine/GNCId.c src/engine/QueryCore.c,
	src/engine/gnc-engine-util.c src/engine/guid.c,
	src/engine/test/test-commodities.c, src/network-utils/gnc-gpg.c:
	Fix signed/unsigned comparison warnings in gcc3.3.


git-svn-id: svn+ssh://svn.gnucash.org/repo/gnucash/branches/gnucash-gnome2-dev@8399 57a11ea4-9604-0410-9ed3-97b8803252fd

+++ src/gnome-utils/gnc-tree-model-example-account.c
@@ -265,7 +265,7 @@ gnc_tree_model_example_account_get_iter (GtkTreeModel *tree_model,
 					 GtkTreePath *path)
 {
 	GNCTreeModelExampleAccount *model = GNC_TREE_MODEL_EXAMPLE_ACCOUNT (tree_model);
-	gint i;
+	guint i;
 
 	g_return_val_if_fail (GNC_IS_TREE_MODEL_EXAMPLE_ACCOUNT (model), FALSE);
 	g_return_val_if_fail (gtk_tree_path_get_depth (path) > 0, FALSE);
@@ -427,7 +427,7 @@ gnc_tree_model_example_account_iter_nth_child (GtkTreeModel *tree_model,
 
 	model = GNC_TREE_MODEL_EXAMPLE_ACCOUNT (tree_model);
 
-	g_return_val_if_fail (n >= 0 && n < g_slist_length (model->priv->accounts), FALSE);
+	g_return_val_if_fail (n >= 0 && n < (int)g_slist_length (model->priv->accounts), FALSE);
 
 	iter->stamp = model->stamp;
 	iter->user_data = g_slist_nth (model->priv->accounts, n);
